+++ tests/test_format_strings.py
@@ -13,7 +13,7 @@ def test_old_format_string(v):
     v.scan("a = 1\n'%(a)s, %(b)d' % locals()")
     check(v.defined_funcs, [])
     check(v.defined_vars, ['a'])
-    check(v.used_vars, ['a', 'b', 'locals'])
+    check(v.used_names, ['a', 'b', 'locals'])
     check(v.unused_vars, [])
 
 
@@ -21,5 +21,5 @@ def test_new_format_string(v):
     v.scan("a = 1\n'{a}, {b}'.format(**locals())")
     check(v.defined_funcs, [])
     check(v.defined_vars, ['a'])
-    check(v.used_vars, ['a', 'b', 'locals'])
+    check(v.used_names, ['a', 'b', 'locals'])
     check(v.unused_vars, [])